requirements.txt
#--- Core Server ---
fastapi
uvicorn

#--- Google AI (ADK Framework) ---
google-adk
google-generativeai
python-dotenv
google-adk

#--- Tools ---
requests
tavily-python

#--- RAG & Memory (Vector Database) ---
faiss-cpu
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-huggingface
sentence-transformers